データ作成部の修正。およびdatagen.py追加。
CreateData::onClick()におけるtoggleGroup.ActiveToggles参照での例外について、toggleGroup.AnyTogglesOnによりトグルの状態を確認することにより回避。また、同関数内のflagを末尾でflag = 0にリセットするようにした。また、CharacterクラスをCharacter.csに分離した。また、JSONデータセット作成用にdatagen.pyを追加。
